refactor: reorganize call code into focused, smaller files
CallController.kt (993 lines) split into:
- CallController.kt (~380 lines) — call orchestration only
- CallMediaManager.kt (~190 lines) — WebRTC factory, EGL, sources,
  tracks, camera lifecycle
- RemoteVideoMonitor.kt (~190 lines) — remote video frame sinks,
  per-peer activity tracking, monitor jobs

CallScreen.kt (1,125 lines) split into:
- CallScreen.kt (~290 lines) — state router + simple screens
- ConnectedCallUI.kt (~320 lines) — active call UI with controls
- CallWidgets.kt (~340 lines) — shared composables (VideoRenderer,
  PeerVideoGrid, GroupCallPictures, GroupCallNames, etc.)
- PipCallUI.kt (~170 lines) — Picture-in-Picture variants

Misplaced files moved:
- ActiveCallHolder → service/call/CallSessionBridge (renamed: it's a
  process-level singleton bridging Activity↔ViewModel, not a "holder")
- CallNotificationReceiver → service/call/ (BroadcastReceiver, not UI)
- showIncomingCallNotification → NotificationUtils (profile picture
  loading doesn't belong in CallController)
