+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/call/CallScreen.kt
@@ -20,7 +20,6 @@
  */
 package com.vitorpamplona.amethyst.ui.call
 
-import android.view.WindowManager
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -29,9 +28,7 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.navigationBars
 import androidx.compose.foundation.layout.padding
@@ -40,66 +37,33 @@ import androidx.compose.foundation.layout.statusBars
 import androidx.compose.foundation.layout.windowInsetsPadding
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.automirrored.filled.VolumeOff
-import androidx.compose.material.icons.automirrored.filled.VolumeUp
-import androidx.compose.material.icons.filled.BluetoothAudio
 import androidx.compose.material.icons.filled.Call
 import androidx.compose.material.icons.filled.CallEnd
-import androidx.compose.material.icons.filled.Mic
-import androidx.compose.material.icons.filled.MicOff
-import androidx.compose.material.icons.filled.PersonAdd
-import androidx.compose.material.icons.filled.Videocam
-import androidx.compose.material.icons.filled.VideocamOff
-import androidx.compose.material.icons.filled.VolumeOff
-import androidx.compose.material.icons.filled.VolumeUp
-import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Snackbar
 import androidx.compose.material3.Text
-import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.DisposableEffect
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableLongStateOf
-import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.compose.ui.viewinterop.AndroidView
 import com.vitorpamplona.amethyst.R
 import com.vitorpamplona.amethyst.commons.call.CallManager
 import com.vitorpamplona.amethyst.commons.call.CallState
-import com.vitorpamplona.amethyst.service.call.AudioRoute
 import com.vitorpamplona.amethyst.service.call.CallController
-import com.vitorpamplona.amethyst.ui.note.BaseUserPicture
-import com.vitorpamplona.amethyst.ui.note.ClickableUserPicture
-import com.vitorpamplona.amethyst.ui.note.UsernameDisplay
-import com.vitorpamplona.amethyst.ui.note.creators.userSuggestions.ShowUserSuggestionList
-import com.vitorpamplona.amethyst.ui.note.creators.userSuggestions.UserSuggestionState
 import com.vitorpamplona.amethyst.ui.screen.loggedIn.AccountViewModel
-import com.vitorpamplona.amethyst.ui.screen.loggedIn.chats.rooms.LoadUser
 import com.vitorpamplona.amethyst.ui.stringRes
-import com.vitorpamplona.quartz.utils.TimeUtils
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
-import org.webrtc.RendererCommon
-import org.webrtc.SurfaceViewRenderer
-import org.webrtc.VideoTrack
 
 @Composable
 fun CallScreen(
@@ -124,8 +88,6 @@ fun CallScreen(
     Box(modifier = Modifier.fillMaxSize()) {
         when (val state = callState) {
             is CallState.Idle -> {
-                // Wait briefly — initiateCall runs async and state may not have
-                // transitioned yet when navigating to this screen
                 LaunchedEffect(Unit) {
                     delay(500)
                     if (callManager.state.value is CallState.Idle) {
@@ -235,8 +197,7 @@ fun CallScreen(
                     action = {
                         Text(
                             stringRes(R.string.call_dismiss),
-                            modifier =
-                                Modifier.padding(8.dp),
+                            modifier = Modifier.padding(8.dp),
                             color = MaterialTheme.colorScheme.inversePrimary,
                         )
                     },
@@ -377,749 +338,6 @@ private fun IncomingCallUI(
                     )
                 }
             }
-        }
-    }
-}
-
-@Composable
-private fun ConnectedCallUI(
-    state: CallState.Connected,
-    callController: CallController?,
-    accountViewModel: AccountViewModel,
-    onHangup: () -> Unit,
-    onToggleMute: () -> Unit,
-    onToggleVideo: () -> Unit,
-    onCycleAudioRoute: () -> Unit,
-    onInvitePeer: (String) -> Unit = {},
-) {
-    var elapsed by remember { mutableLongStateOf(0L) }
-
-    LaunchedEffect(state.startedAtEpoch) {
-        while (true) {
-            elapsed = TimeUtils.now() - state.startedAtEpoch
-            delay(1000)
-        }
-    }
-
-    val emptyVideoFlow = remember { kotlinx.coroutines.flow.MutableStateFlow<VideoTrack?>(null) }
-    val emptyTracksFlow = remember { kotlinx.coroutines.flow.MutableStateFlow<Map<String, VideoTrack>>(emptyMap()) }
-    val emptySetFlow = remember { kotlinx.coroutines.flow.MutableStateFlow<Set<String>>(emptySet()) }
-    val remoteVideoTracks by (callController?.remoteVideoTracks ?: emptyTracksFlow).collectAsState()
-    val activePeerVideos by (callController?.activePeerVideos ?: emptySetFlow).collectAsState()
-    val localVideoTrack by (callController?.localVideoTrack ?: emptyVideoFlow).collectAsState()
-    val defaultFalse = remember { kotlinx.coroutines.flow.MutableStateFlow(false) }
-    val defaultTrue = remember { kotlinx.coroutines.flow.MutableStateFlow(true) }
-    val isRemoteVideoActive by (callController?.isRemoteVideoActive ?: defaultFalse).collectAsState()
-    val defaultRoute = remember { kotlinx.coroutines.flow.MutableStateFlow(AudioRoute.EARPIECE) }
-    val isAudioMuted by (callController?.isAudioMuted ?: defaultFalse).collectAsState()
-    val isVideoEnabled by (callController?.isVideoEnabled ?: defaultTrue).collectAsState()
-    val currentAudioRoute by (callController?.audioRoute ?: defaultRoute).collectAsState()
-    val hasActiveVideo =
-        state.callType == com.vitorpamplona.quartz.nipACWebRtcCalls.tags.CallType.VIDEO ||
-            isVideoEnabled ||
-            remoteVideoTracks.isNotEmpty()
-
-    var showAddParticipant by remember { mutableStateOf(false) }
-
-    if (showAddParticipant) {
-        AddParticipantDialog(
-            accountViewModel = accountViewModel,
-            existingPeers = state.allPeerPubKeys,
-            onInvite = { peerPubKey ->
-                onInvitePeer(peerPubKey)
-                showAddParticipant = false
-            },
-            onDismiss = { showAddParticipant = false },
-        )
-    }
-
-    Box(
-        modifier =
-            Modifier
-                .fillMaxSize()
-                .background(Color.Black),
-    ) {
-        val otherMembers =
-            remember(state.allPeerPubKeys) {
-                state.allPeerPubKeys - accountViewModel.account.signer.pubKey
-            }
-
-        if (hasActiveVideo) {
-            // Video call: always show the peer grid with video for active peers, avatar for inactive
-            PeerVideoGrid(
-                peerPubKeys = otherMembers,
-                remoteVideoTracks = remoteVideoTracks,
-                activePeerVideos = activePeerVideos,
-                eglBase = callController?.getEglBase(),
-                accountViewModel = accountViewModel,
-                modifier = Modifier.fillMaxSize(),
-            )
-
-            // Local video (small pip in corner) — only when camera is active
-            if (isVideoEnabled) {
-                localVideoTrack?.let { track ->
-                    VideoRenderer(
-                        videoTrack = track,
-                        eglBase = callController?.getEglBase(),
-                        modifier =
-                            Modifier
-                                .size(120.dp, 160.dp)
-                                .align(Alignment.TopEnd)
-                                .windowInsetsPadding(WindowInsets.statusBars)
-                                .padding(16.dp),
-                        mirror = true,
-                    )
-                }
-            }
-
-            // Timer overlay
-            Text(
-                text = formatDuration(elapsed),
-                color = Color.White.copy(alpha = 0.7f),
-                fontSize = 14.sp,
-                modifier =
-                    Modifier
-                        .align(Alignment.TopCenter)
-                        .windowInsetsPadding(WindowInsets.statusBars)
-                        .padding(top = 16.dp),
-            )
-
-            if (state.pendingPeerPubKeys.isNotEmpty()) {
-                Text(
-                    text = stringRes(R.string.call_waiting_for_others),
-                    color = Color.White.copy(alpha = 0.5f),
-                    fontSize = 13.sp,
-                    modifier =
-                        Modifier
-                            .align(Alignment.TopCenter)
-                            .windowInsetsPadding(WindowInsets.statusBars)
-                            .padding(top = 38.dp),
-                )
-            }
-        } else {
-            // Voice call: show avatars and names
-            Column(
-                modifier = Modifier.fillMaxSize(),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Center,
-            ) {
-                GroupCallPictures(
-                    peerPubKeys = otherMembers,
-                    size = 120.dp,
-                    accountViewModel = accountViewModel,
-                )
-                Spacer(modifier = Modifier.height(16.dp))
-                GroupCallNames(
-                    peerPubKeys = otherMembers,
-                    accountViewModel = accountViewModel,
-                    textColor = Color.White,
-                )
-                Spacer(modifier = Modifier.height(8.dp))
-                Text(
-                    text = formatDuration(elapsed),
-                    color = Color.White.copy(alpha = 0.7f),
-                    fontSize = 16.sp,
-                )
-                if (state.pendingPeerPubKeys.isNotEmpty()) {
-                    Spacer(modifier = Modifier.height(4.dp))
-                    Text(
-                        text = stringRes(R.string.call_waiting_for_others),
-                        color = Color.White.copy(alpha = 0.5f),
-                        fontSize = 13.sp,
-                    )
-                }
-            }
-        }
-
-        // Controls at bottom
-        Column(
-            modifier =
-                Modifier
-                    .align(Alignment.BottomCenter)
-                    .windowInsetsPadding(WindowInsets.navigationBars)
-                    .padding(bottom = 24.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                horizontalArrangement = Arrangement.SpaceEvenly,
-            ) {
-                IconButton(
-                    onClick = onToggleMute,
-                    modifier = Modifier.size(56.dp),
-                ) {
-                    Icon(
-                        imageVector = if (isAudioMuted) Icons.Default.MicOff else Icons.Default.Mic,
-                        contentDescription = stringRes(if (isAudioMuted) R.string.call_unmute else R.string.call_mute),
-                        tint = if (isAudioMuted) Color.Red else Color.White,
-                        modifier = Modifier.size(28.dp),
-                    )
-                }
-                IconButton(
-                    onClick = onToggleVideo,
-                    modifier = Modifier.size(56.dp),
-                ) {
-                    Icon(
-                        imageVector = if (isVideoEnabled) Icons.Default.Videocam else Icons.Default.VideocamOff,
-                        contentDescription = stringRes(if (isVideoEnabled) R.string.call_camera_off else R.string.call_camera_on),
-                        tint = if (!isVideoEnabled) Color.Red else Color.White,
-                        modifier = Modifier.size(28.dp),
-                    )
-                }
-                IconButton(
-                    onClick = onCycleAudioRoute,
-                    modifier = Modifier.size(56.dp),
-                ) {
-                    Icon(
-                        imageVector =
-                            when (currentAudioRoute) {
-                                AudioRoute.EARPIECE -> Icons.AutoMirrored.Filled.VolumeOff
-                                AudioRoute.SPEAKER -> Icons.AutoMirrored.Filled.VolumeUp
-                                AudioRoute.BLUETOOTH -> Icons.Default.BluetoothAudio
-                            },
-                        contentDescription =
-                            stringRes(
-                                when (currentAudioRoute) {
-                                    AudioRoute.EARPIECE -> R.string.call_earpiece
-                                    AudioRoute.SPEAKER -> R.string.call_speaker
-                                    AudioRoute.BLUETOOTH -> R.string.call_bluetooth
-                                },
-                            ),
-                        tint =
-                            when (currentAudioRoute) {
-                                AudioRoute.EARPIECE -> Color.White
-                                AudioRoute.SPEAKER -> Color.Cyan
-                                AudioRoute.BLUETOOTH -> Color(0xFF2196F3)
-                            },
-                        modifier = Modifier.size(28.dp),
-                    )
-                }
-                IconButton(
-                    onClick = { showAddParticipant = true },
-                    modifier = Modifier.size(56.dp),
-                ) {
-                    Icon(
-                        imageVector = Icons.Default.PersonAdd,
-                        contentDescription = stringRes(R.string.call_add_participant),
-                        tint = Color.White,
-                        modifier = Modifier.size(28.dp),
-                    )
-                }
-            }
-            Spacer(modifier = Modifier.height(24.dp))
-            FloatingActionButton(
-                onClick = onHangup,
-                containerColor = Color.Red,
-                shape = CircleShape,
-                modifier = Modifier.size(64.dp),
-            ) {
-                Icon(
-                    Icons.Default.CallEnd,
-                    contentDescription = stringRes(R.string.call_hangup),
-                    tint = Color.White,
-                    modifier = Modifier.size(32.dp),
-                )
-            }
-        }
-    }
-}
-
-@Composable
-private fun PeerVideoGrid(
-    peerPubKeys: Set<String>,
-    remoteVideoTracks: Map<String, VideoTrack>,
-    activePeerVideos: Set<String>,
-    eglBase: org.webrtc.EglBase?,
-    accountViewModel: AccountViewModel,
-    modifier: Modifier = Modifier,
-) {
-    val peers = remember(peerPubKeys) { peerPubKeys.toList() }
-
-    if (peers.size == 1) {
-        val peerKey = peers[0]
-        val track = remoteVideoTracks[peerKey]
-        if (track != null && peerKey in activePeerVideos) {
-            VideoRenderer(
-                videoTrack = track,
-                eglBase = eglBase,
-                modifier = modifier,
-                mirror = false,
-            )
-        } else {
-            PeerAvatarCell(
-                peerPubKey = peerKey,
-                accountViewModel = accountViewModel,
-                modifier = modifier,
-            )
-        }
-    } else {
-        val columns =
-            when {
-                peers.size <= 2 -> 1
-                else -> 2
-            }
-
-        Column(modifier = modifier) {
-            peers.chunked(columns).forEach { row ->
-                Row(
-                    modifier = Modifier.weight(1f).fillMaxWidth(),
-                ) {
-                    row.forEach { peerKey ->
-                        val track = remoteVideoTracks[peerKey]
-                        if (track != null && peerKey in activePeerVideos) {
-                            VideoRenderer(
-                                videoTrack = track,
-                                eglBase = eglBase,
-                                modifier = Modifier.weight(1f).fillMaxHeight(),
-                                mirror = false,
-                            )
-                        } else {
-                            PeerAvatarCell(
-                                peerPubKey = peerKey,
-                                accountViewModel = accountViewModel,
-                                modifier = Modifier.weight(1f).fillMaxHeight(),
-                            )
-                        }
-                    }
-                    repeat(columns - row.size) {
-                        Spacer(modifier = Modifier.weight(1f))
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun PeerAvatarCell(
-    peerPubKey: String,
-    accountViewModel: AccountViewModel,
-    modifier: Modifier = Modifier,
-) {
-    Box(
-        modifier = modifier.background(Color.DarkGray),
-        contentAlignment = Alignment.Center,
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-        ) {
-            LoadUser(baseUserHex = peerPubKey, accountViewModel = accountViewModel) { user ->
-                if (user != null) {
-                    ClickableUserPicture(
-                        baseUser = user,
-                        size = 80.dp,
-                        accountViewModel = accountViewModel,
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                    UsernameDisplay(
-                        baseUser = user,
-                        accountViewModel = accountViewModel,
-                        fontWeight = FontWeight.Bold,
-                        textColor = Color.White,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun VideoRenderer(
-    videoTrack: VideoTrack,
-    eglBase: org.webrtc.EglBase?,
-    modifier: Modifier = Modifier,
-    mirror: Boolean = false,
-) {
-    AndroidView(
-        modifier = modifier,
-        factory = { ctx ->
-            SurfaceViewRenderer(ctx).apply {
-                setMirror(mirror)
-                setScalingType(RendererCommon.ScalingType.SCALE_ASPECT_FILL)
-                eglBase?.eglBaseContext?.let { init(it, null) }
-                videoTrack.addSink(this)
-            }
-        },
-        onRelease = { renderer ->
-            videoTrack.removeSink(renderer)
-            renderer.release()
-        },
-    )
-}
-
-private fun formatDuration(seconds: Long): String {
-    val mins = seconds / 60
-    val secs = seconds % 60
-    return "%02d:%02d".format(mins, secs)
-}
-
-@Composable
-private fun PipCallUI(
-    peerPubKeys: Set<String>,
-    statusText: String,
-    accountViewModel: AccountViewModel,
-) {
-    Box(
-        modifier =
-            Modifier
-                .fillMaxSize()
-                .background(MaterialTheme.colorScheme.surface),
-        contentAlignment = Alignment.Center,
-    ) {
-        Column(
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center,
-        ) {
-            GroupCallPictures(
-                peerPubKeys = peerPubKeys,
-                size = 48.dp,
-                accountViewModel = accountViewModel,
-            )
-            Spacer(modifier = Modifier.height(4.dp))
-            Text(
-                text = statusText,
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                fontSize = 10.sp,
-            )
-        }
-    }
-}
-
-@Composable
-private fun PipConnectedCallUI(
-    state: CallState.Connected,
-    callController: CallController?,
-    accountViewModel: AccountViewModel,
-) {
-    var elapsed by remember { mutableLongStateOf(0L) }
-
-    LaunchedEffect(state.startedAtEpoch) {
-        while (true) {
-            elapsed = TimeUtils.now() - state.startedAtEpoch
-            delay(1000)
-        }
-    }
-
-    val emptyTracksFlow = remember { kotlinx.coroutines.flow.MutableStateFlow<Map<String, VideoTrack>>(emptyMap()) }
-    val emptySetFlow = remember { kotlinx.coroutines.flow.MutableStateFlow<Set<String>>(emptySet()) }
-    val remoteVideoTracks by (callController?.remoteVideoTracks ?: emptyTracksFlow).collectAsState()
-    val activePeerVideos by (callController?.activePeerVideos ?: emptySetFlow).collectAsState()
-    val defaultFalse = remember { kotlinx.coroutines.flow.MutableStateFlow(false) }
-    val isRemoteVideoActive by (callController?.isRemoteVideoActive ?: defaultFalse).collectAsState()
-    val isVideoEnabled by (callController?.isVideoEnabled ?: defaultFalse).collectAsState()
-    val hasActiveVideo =
-        state.callType == com.vitorpamplona.quartz.nipACWebRtcCalls.tags.CallType.VIDEO ||
-            isVideoEnabled ||
-            remoteVideoTracks.isNotEmpty()
-
-    val otherMembers =
-        remember(state.allPeerPubKeys) {
-            state.allPeerPubKeys - accountViewModel.account.signer.pubKey
-        }
-
-    Box(
-        modifier =
-            Modifier
-                .fillMaxSize()
-                .background(Color.Black),
-    ) {
-        if (hasActiveVideo) {
-            // Video active: show first active peer's video or avatar
-            val firstActivePeer = otherMembers.firstOrNull { it in activePeerVideos }
-            val activeTrack = firstActivePeer?.let { remoteVideoTracks[it] }
-            if (activeTrack != null) {
-                VideoRenderer(
-                    videoTrack = activeTrack,
-                    eglBase = callController?.getEglBase(),
-                    modifier = Modifier.fillMaxSize(),
-                    mirror = false,
-                )
-            } else {
-                Column(
-                    modifier = Modifier.fillMaxSize(),
-                    horizontalAlignment = Alignment.CenterHorizontally,
-                    verticalArrangement = Arrangement.Center,
-                ) {
-                    GroupCallPictures(
-                        peerPubKeys = otherMembers,
-                        size = 48.dp,
-                        accountViewModel = accountViewModel,
-                    )
-                }
-            }
-        } else if (!isRemoteVideoActive) {
-            Column(
-                modifier = Modifier.fillMaxSize(),
-                horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.Center,
-            ) {
-                GroupCallPictures(
-                    peerPubKeys = otherMembers,
-                    size = 48.dp,
-                    accountViewModel = accountViewModel,
-                )
-                Spacer(modifier = Modifier.height(4.dp))
-                Text(
-                    text = formatDuration(elapsed),
-                    color = Color.White.copy(alpha = 0.7f),
-                    fontSize = 10.sp,
-                )
-            }
-        }
-
-        // Timer overlay
-        Text(
-            text = formatDuration(elapsed),
-            color = Color.White.copy(alpha = 0.7f),
-            fontSize = 10.sp,
-            modifier =
-                Modifier
-                    .align(Alignment.TopCenter)
-                    .padding(top = 4.dp),
-        )
-    }
-}
-
-@Composable
-private fun AddParticipantDialog(
-    accountViewModel: AccountViewModel,
-    existingPeers: Set<String>,
-    onInvite: (String) -> Unit,
-    onDismiss: () -> Unit,
-) {
-    val userSuggestions =
-        remember {
-            UserSuggestionState(accountViewModel.account, accountViewModel.nip05ClientBuilder())
-        }
-    var searchText by remember { mutableStateOf("") }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text(stringRes(R.string.call_add_participant)) },
-        text = {
-            Column {
-                OutlinedTextField(
-                    value = searchText,
-                    onValueChange = {
-                        searchText = it
-                        userSuggestions.processCurrentWord(it)
-                    },
-                    label = { Text(stringRes(R.string.call_search_users)) },
-                    singleLine = true,
-                    modifier = Modifier.fillMaxWidth(),
-                )
-                Box(modifier = Modifier.height(300.dp)) {
-                    ShowUserSuggestionList(
-                        userSuggestions = userSuggestions,
-                        onSelect = { user ->
-                            if (user.pubkeyHex !in existingPeers) {
-                                onInvite(user.pubkeyHex)
-                            }
-                        },
-                        accountViewModel = accountViewModel,
-                    )
-                }
-            }
-        },
-        confirmButton = {},
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text(stringRes(R.string.call_dismiss))
-            }
-        },
-    )
-}
-
-@Composable
-private fun GroupCallPictures(
-    peerPubKeys: Set<String>,
-    size: Dp,
-    accountViewModel: AccountViewModel,
-) {
-    val userList = remember(peerPubKeys) { peerPubKeys.toList() }
-    val displayCount = minOf(userList.size, 4)
-    val remaining = userList.size - displayCount
-
-    when (userList.size) {
-        0 -> {}
-
-        1 -> {
-            LoadUser(baseUserHex = userList[0], accountViewModel = accountViewModel) { user ->
-                if (user != null) {
-                    ClickableUserPicture(
-                        baseUser = user,
-                        size = size,
-                        accountViewModel = accountViewModel,
-                    )
-                }
-            }
-        }
-
-        else -> {
-            Box(Modifier.size(size), contentAlignment = Alignment.TopEnd) {
-                when (displayCount) {
-                    2 -> {
-                        BaseUserPicture(
-                            baseUserHex = userList[0],
-                            size = size.div(1.5f),
-                            accountViewModel = accountViewModel,
-                            outerModifier = Modifier.size(size.div(1.5f)).align(Alignment.CenterStart),
-                        )
-                        BaseUserPicture(
-                            baseUserHex = userList[1],
-                            size = size.div(1.5f),
-                            accountViewModel = accountViewModel,
-                            outerModifier = Modifier.size(size.div(1.5f)).align(Alignment.CenterEnd),
-                        )
-                    }
-
-                    3 -> {
-                        BaseUserPicture(
-                            baseUserHex = userList[0],
-                            size = size.div(1.8f),
-                            accountViewModel = accountViewModel,
-                            outerModifier = Modifier.size(size.div(1.8f)).align(Alignment.BottomStart),
-                        )
-                        BaseUserPicture(
-                            baseUserHex = userList[1],
-                            size = size.div(1.8f),
-                            accountViewModel = accountViewModel,
-                            outerModifier = Modifier.size(size.div(1.8f)).align(Alignment.TopCenter),
-                        )
-                        BaseUserPicture(
-                            baseUserHex = userList[2],
-                            size = size.div(1.8f),
-                            accountViewModel = accountViewModel,
-                            outerModifier = Modifier.size(size.div(1.8f)).align(Alignment.BottomEnd),
-                        )
-                    }
-
-                    else -> {
-                        BaseUserPicture(
-                            baseUserHex = userList[0],
-                            size = size.div(2f),
-                            accountViewModel = accountViewModel,
-                            outerModifier = Modifier.size(size.div(2f)).align(Alignment.BottomStart),
-                        )
-                        BaseUserPicture(
-                            baseUserHex = userList[1],
-                            size = size.div(2f),
-                            accountViewModel = accountViewModel,
-                            outerModifier = Modifier.size(size.div(2f)).align(Alignment.TopStart),
-                        )
-                        BaseUserPicture(
-                            baseUserHex = userList[2],
-                            size = size.div(2f),
-                            accountViewModel = accountViewModel,
-                            outerModifier = Modifier.size(size.div(2f)).align(Alignment.BottomEnd),
-                        )
-                        if (remaining > 0) {
-                            Box(
-                                modifier =
-                                    Modifier
-                                        .size(size.div(2f))
-                                        .align(Alignment.TopEnd)
-                                        .background(
-                                            MaterialTheme.colorScheme.surfaceVariant,
-                                            CircleShape,
-                                        ),
-                                contentAlignment = Alignment.Center,
-                            ) {
-                                Text(
-                                    text = "+$remaining",
-                                    color = MaterialTheme.colorScheme.onSurfaceVariant,
-                                    fontSize = (size.value / 5).sp,
-                                    fontWeight = FontWeight.Bold,
-                                )
-                            }
-                        } else {
-                            BaseUserPicture(
-                                baseUserHex = userList[3],
-                                size = size.div(2f),
-                                accountViewModel = accountViewModel,
-                                outerModifier = Modifier.size(size.div(2f)).align(Alignment.TopEnd),
-                            )
-                        }
-                    }
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun GroupCallNames(
-    peerPubKeys: Set<String>,
-    accountViewModel: AccountViewModel,
-    textColor: Color = MaterialTheme.colorScheme.onSurface,
-) {
-    val userList = remember(peerPubKeys) { peerPubKeys.toList() }
-
-    when (userList.size) {
-        0 -> {}
-
-        1 -> {
-            LoadUser(baseUserHex = userList[0], accountViewModel = accountViewModel) { user ->
-                if (user != null) {
-                    UsernameDisplay(
-                        baseUser = user,
-                        accountViewModel = accountViewModel,
-                        fontWeight = FontWeight.Bold,
-                        textColor = textColor,
-                    )
-                }
-            }
-        }
-
-        else -> {
-            val displayCount = minOf(userList.size, 2)
-            val remaining = userList.size - displayCount
-
-            Row(
-                horizontalArrangement = Arrangement.Center,
-                verticalAlignment = Alignment.CenterVertically,
-            ) {
-                for (i in 0 until displayCount) {
-                    if (i > 0) {
-                        Text(
-                            text = ", ",
-                            fontWeight = FontWeight.Bold,
-                            color = textColor,
-                        )
-                    }
-                    LoadUser(baseUserHex = userList[i], accountViewModel = accountViewModel) { user ->
-                        if (user != null) {
-                            UsernameDisplay(
-                                baseUser = user,
-                                accountViewModel = accountViewModel,
-                                fontWeight = FontWeight.Bold,
-                                textColor = textColor,
-                            )
-                        }
-                    }
-                }
-                if (remaining > 0) {
-                    Text(
-                        text = " +$remaining",
-                        fontWeight = FontWeight.Bold,
-                        color = textColor,
-                        textAlign = TextAlign.Center,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
-private fun KeepScreenOn() {
-    val context = LocalContext.current
-    DisposableEffect(Unit) {
-        val window = (context as? android.app.Activity)?.window
-        window?.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-        onDispose {
-            window?.clearFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
         }
     }
 }
